+++ test.py
@@ -4,8 +4,7 @@ import numpy as np
 
 from src import isovec as iso
 
-# TODO: copy function
-# TODO (far): IAEA Livechart Data Download API class
 
+# TODO (far): IAEA Livechart Data Download API class
 
 pass